moving video player to new file
